pyproject.toml
[project]
name = "vggt-jax"
version = "0.1.0"
description = "VGGT implementation in JAX + Flax NNX"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "jax[cuda13]>=0.9.0.1",
  "flax>=0.12.4",
  "orbax-checkpoint>=0.11.0",
  "einops>=0.8.0",
  "huggingface-hub>=0.30.0",
  "numpy>=2.0,<3.0",
  "Pillow>=10.0.0",
  "safetensors>=0.5.0",
]

[project.optional-dependencies]
data = [
  "grain==0.2.15",
]
demo = [
  "gradio>=5.0.0",
  "viser>=0.2.0",
  "trimesh>=4.0.0",
  "opencv-python>=4.10.0",
  "onnxruntime>=1.18.0",
  "tqdm>=4.66.0",
  "pycolmap>=3.10.0",
]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.11.0",
]
convert = [
  "safetensors>=0.5.0",
]
track = [
  "pycolmap>=3.10.0",
  "opencv-python>=4.10.0",
]
train = [
  "hydra-core>=1.3.0",
  "omegaconf>=2.3.0",
  "iopath>=0.1.10",
  "wcmatch>=10.0",
  "tensorboard>=2.17.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["vggt*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
